LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_removed_mouse-expected.txt

child
done

FAIL pointer events from mouse received before/after child removal at pointerdown assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerover@parent,pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerup@parent,pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
FAIL pointer events from mouse received before/after child removal at pointerup assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerover@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent"
FAIL mouse events from mouse received before/after child removal at mousedown assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseover@parent,mouseup@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseup@parent,mouseover@parent,mouseenter@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent"
FAIL mouse events from mouse received before/after child removal at mouseup assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,(child-removed),mouseover@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,(child-removed),mouseover@parent,mouseenter@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent"
